Restructure to use src-tree layout
