+++ src/snappy_device_agents/data/pi-desktop/preseed.cfg
