+++ src/snappy_device_agents/devices/maas2/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2017-2019 Canonical
+# Copyright (C) 2017-2023 Canonical
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -12,17 +12,24 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Ubuntu Raspberry PI CM3 support code."""
+"""Ubuntu MaaS 2.x CLI support code."""
 
 import logging
+
 import yaml
 
 import snappy_device_agents
-from devices.cm3.cm3 import CM3
 from snappy_device_agents import logmsg
-from devices import catch, DefaultDevice, RecoveryError, SerialLogger
+from snappy_device_agents.devices import (
+    DefaultDevice,
+    ProvisioningError,
+    RecoveryError,
+    SerialLogger,
+    catch,
+)
+from snappy_device_agents.devices.maas2.maas2 import Maas2
 
-device_name = "cm3"
+device_name = "maas2"
 
 
 class DeviceAgent(DefaultDevice):
@@ -35,7 +42,7 @@ class DeviceAgent(DefaultDevice):
         with open(args.config) as configfile:
             config = yaml.safe_load(configfile)
         snappy_device_agents.configure_logging(config)
-        device = CM3(args.config, args.job_data)
+        device = Maas2(args.config, args.job_data)
         logmsg(logging.INFO, "BEGIN provision")
         logmsg(logging.INFO, "Provisioning device")
         serial_host = config.get("serial_host")
@@ -46,6 +53,9 @@ class DeviceAgent(DefaultDevice):
         serial_proc.start()
         try:
             device.provision()
+        except ProvisioningError as e:
+            logmsg(logging.ERROR, "Provisioning failed: {}".format(str(e)))
+            return 1
         except Exception as e:
             raise e
         finally:
