+++ src/snappy_device_agents/devices/maas2/maas2.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2017 Canonical
+# Copyright (C) 2023 Canonical
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -22,7 +22,8 @@ import time
 from collections import OrderedDict
 
 import yaml
-from devices import ProvisioningError, RecoveryError
+
+from snappy_device_agents.devices import ProvisioningError, RecoveryError
 
 logger = logging.getLogger()
