+++ src/snappy_device_agents/devices/muxpi/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2016-2019 Canonical
+# Copyright (C) 2017-2023 Canonical
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -12,17 +12,23 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Dragonboard support code."""
+"""Ubuntu Raspberry PI muxpi support code."""
 
 import logging
+
 import yaml
 
 import snappy_device_agents
-from devices.dragonboard.dragonboard import Dragonboard
 from snappy_device_agents import logmsg
-from devices import catch, DefaultDevice, RecoveryError, SerialLogger
+from snappy_device_agents.devices import (
+    DefaultDevice,
+    RecoveryError,
+    SerialLogger,
+    catch,
+)
+from snappy_device_agents.devices.muxpi.muxpi import MuxPi
 
-device_name = "dragonboard"
+device_name = "muxpi"
 
 
 class DeviceAgent(DefaultDevice):
@@ -35,9 +41,9 @@ class DeviceAgent(DefaultDevice):
         with open(args.config) as configfile:
             config = yaml.safe_load(configfile)
         snappy_device_agents.configure_logging(config)
-        device = Dragonboard(args.config, args.job_data)
+        device = MuxPi(args.config, args.job_data)
         logmsg(logging.INFO, "BEGIN provision")
-        logmsg(logging.INFO, "Booting Master Image")
+        logmsg(logging.INFO, "Provisioning device")
         serial_host = config.get("serial_host")
         serial_port = config.get("serial_port")
         serial_proc = SerialLogger(
@@ -50,3 +56,4 @@ class DeviceAgent(DefaultDevice):
             raise e
         finally:
             serial_proc.stop()
+        logmsg(logging.INFO, "END provision")
