+++ src/snappy_device_agents/devices/muxpi/muxpi.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2017-2020 Canonical
+# Copyright (C) 2017-2023 Canonical
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -20,13 +20,12 @@ import multiprocessing
 import os
 import subprocess
 import time
+from contextlib import contextmanager
+
 import yaml
 
 import snappy_device_agents
-
-from contextlib import contextmanager
-
-from devices import ProvisioningError, RecoveryError
+from snappy_device_agents.devices import ProvisioningError, RecoveryError
 
 logger = logging.getLogger()
