+++ src/snappy_device_agents/devices/oemrecovery/oemrecovery.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2018 Canonical
+# Copyright (C) 2023 Canonical
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -18,10 +18,11 @@ import json
 import logging
 import subprocess
 import time
+
 import yaml
 
-from devices import ProvisioningError, RecoveryError
 from snappy_device_agents import CmdTimeoutError
+from snappy_device_agents.devices import ProvisioningError, RecoveryError
 
 logger = logging.getLogger()
