+++ tox.ini
@@ -10,8 +10,8 @@ deps =
     pylint
     pytest-cov
 commands =
-    {envbindir}/python setup.py develop
-    {envbindir}/python -m black --check setup.py snappy_device_agents devices tests
-    {envbindir}/python -m flake8 setup.py snappy_device_agents devices
-    #{envbindir}/python -m pylint snappy_device_agents devices
-    {envbindir}/python -m pytest --doctest-modules --cov=snappy_device_agents --cov=devices
+    {envbindir}/pip3 install .
+    {envbindir}/python -m black --check src
+    {envbindir}/python -m flake8 src
+    #{envbindir}/python -m pylint src
+    {envbindir}/python -m pytest --doctest-modules --cov=src
